refactor/split to different files, cleanup and add more documentation comments
