+++ src/vs/workbench/contrib/chat/common/promptSyntax/parsers/promptHeader/metadata/description.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { PromptStringMetadata } from './record.js';
+import { PromptStringMetadata } from './base/string.js';
 import { FrontMatterRecord, FrontMatterToken } from '../../../../../../../../editor/common/codecs/frontMatterCodec/tokens/index.js';
 
 /**
