+++ src/vs/workbench/contrib/chat/common/promptSyntax/parsers/promptHeader/metadata/mode.ts
@@ -3,8 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { PromptEnumMetadata } from './record.js';
 import { ChatMode } from '../../../../constants.js';
+import { PromptEnumMetadata } from './base/enum.js';
 import { FrontMatterRecord, FrontMatterToken } from '../../../../../../../../editor/common/codecs/frontMatterCodec/tokens/index.js';
 
 /**
